chore: Move LazyImage and rogressBar to atoms
